--- CONTRIBUTING.md
@@ -1,31 +1,0 @@
-# Contributing to AppDynamics Extensions
-
-If you would like to contribute code you can do so through GitHub by forking the repository and sending a pull request (on a branch other than `master` or `gh-pages`).
-
-When submitting code, please make every effort to follow existing conventions and style in order to keep the code as consistent as possible.
-
-## License
-
-By contributing your code, you agree to license your contribution under the terms of the APLv2: https://github.com/extensions-commons/blob/master/LICENSE
-
-All files are released with the Apache 2.0 license.
-
-If you are adding a new file it should include a copyright header:
-
-```
-/**
- * Copyright 2013 AppDynamics, Inc.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
- ```
