--- src/main/java/com/appdynamics/monitors/nginx/Config/MetricConverter.java
@@ -1,38 +1,0 @@
-/*
- *
- *   Copyright 2018. AppDynamics LLC and its affiliates.
- *   All Rights Reserved.
- *   This is unpublished proprietary source code of AppDynamics LLC and its affiliates.
- *   The copyright notice above does not evidence any actual or intended publication of such source code.
- *
- */
-
-package com.appdynamics.monitors.nginx.Config;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
-
-@XmlAccessorType(XmlAccessType.FIELD)
-public class MetricConverter {
-    @XmlAttribute(name = "str")
-    private String label;
-    @XmlAttribute(name = "value")
-    private String value;
-
-    public String getLabel() {
-        return label;
-    }
-
-    public void setLabel(String label) {
-        this.label = label;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
-}
