--- src/main/java/com/appdynamics/monitors/nginx/Config/Stat.java
@@ -1,81 +1,0 @@
-package com.appdynamics.monitors.nginx.Config;
-/*
- * Copyright 2018. AppDynamics LLC and its affiliates.
- * All Rights Reserved.
- * This is unpublished proprietary source code of AppDynamics LLC and its affiliates.
- * The copyright notice above does not evidence any actual or intended publication of such source code.
- */
-
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
-
-@XmlAccessorType(XmlAccessType.FIELD)
-public class Stat {
-    @XmlAttribute
-    private String suburl;
-    @XmlAttribute
-    private String name;
-    @XmlElement(name = "metric")
-    private MetricConfig[] metricConfig;
-    @XmlElement(name = "stat")
-    public Stat[] stats;
-
-    public MetricConfig[] getMetricConfig() {
-        return metricConfig;
-    }
-
-    public void setMetricConfig(MetricConfig[] metricConfig) {
-        this.metricConfig = metricConfig;
-    }
-
-    public String getSubUrl() {
-        return suburl;
-    }
-
-    public void setSubUrl(String suburl) {
-        this.suburl = suburl;
-    }
-
-    public Stat[] getStats() {
-        return stats;
-    }
-
-    public void setStats(Stat[] stats) {
-        this.stats = stats;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @XmlRootElement
-    @XmlAccessorType(XmlAccessType.FIELD)
-    public static class Stats {
-        @XmlAttribute
-        private String url;
-        @XmlElement(name = "stat")
-        private Stat[] stats;
-
-        public Stat[] getStats() {
-            return stats;
-        }
-
-        public void setStats(Stat[] stats) {
-            this.stats = stats;
-        }
-        public String getUrl(){
-            return url;
-        }
-        public void setUrl(String url){
-            this.url = url;
-        }
-    }
-}
