--- src/main/resources/conf/monitor.xml
@@ -1,33 +1,0 @@
-<!--
-  ~ Copyright 2013. AppDynamics LLC and its affiliates.
-  ~  All Rights Reserved.
-  ~  This is unpublished proprietary source code of AppDynamics LLC and its affiliates.
-  ~  The copyright notice above does not evidence any actual or intended publication of such source code.
-  -->
-
-<monitor>
-    <name>NGinXMonitor</name>
-    <type>managed</type>
-    <description>NGinX server monitor</description>
-    <monitor-configuration></monitor-configuration>
-    <monitor-run-task>
-        <execution-style>periodic</execution-style>
-        <execution-frequency-in-seconds>60</execution-frequency-in-seconds>
-        <name>NGinX Monitor Run Task</name>
-        <display-name>NGinX Monitor Task</display-name>
-        <description>NGinX Monitor Task</description>
-        <type>java</type>
-        <execution-timeout-in-secs>60</execution-timeout-in-secs>
-        <task-arguments>
-            <argument name="config-file" is-required="true" default-value="monitors/NGinXMonitor/config.yml" />
-            <argument name="metric-file" is-required="true" default-value="monitors/NGinXMonitor/metrics.xml" />
-            <!-- Optional Properties  -->
-            <!--<argument name="metric-prefix" is-required="false" default-value="Custom Metrics|WebServer|NGinX"/>-->
-
-        </task-arguments>
-        <java-task>
-            <classpath>nginx-monitoring-extension.jar</classpath>
-            <impl-class>com.appdynamics.monitors.nginx.NGinXMonitor</impl-class>
-        </java-task>
-    </monitor-run-task>
-</monitor>
