+++ src/main/resources/monitor.xml
@@ -1,0 +1,33 @@
+<!--
+  ~ Copyright 2013. AppDynamics LLC and its affiliates.
+  ~  All Rights Reserved.
+  ~  This is unpublished proprietary source code of AppDynamics LLC and its affiliates.
+  ~  The copyright notice above does not evidence any actual or intended publication of such source code.
+  -->
+
+<monitor>
+    <name>NginxMonitor</name>
+    <type>managed</type>
+    <description>Nginx server monitor</description>
+    <monitor-configuration></monitor-configuration>
+    <monitor-run-task>
+        <execution-style>periodic</execution-style>
+        <execution-frequency-in-seconds>60</execution-frequency-in-seconds>
+        <name>Nginx Monitor Run Task</name>
+        <display-name>Nginx Monitor Task</display-name>
+        <description>Nginx Monitor Task</description>
+        <type>java</type>
+        <execution-timeout-in-secs>60</execution-timeout-in-secs>
+        <task-arguments>
+            <argument name="config-file" is-required="true" default-value="monitors/NginxMonitor/config.yml" />
+            <argument name="metric-file" is-required="true" default-value="monitors/NginxMonitor/metrics.xml" />
+            <!-- Optional Properties  -->
+            <!--<argument name="metric-prefix" is-required="false" default-value="Custom Metrics|WebServer|NGinX"/>-->
+
+        </task-arguments>
+        <java-task>
+            <classpath>nginx-monitoring-extension.jar</classpath>
+            <impl-class>com.appdynamics.extensions.nginx.NginxMonitor</impl-class>
+        </java-task>
+    </monitor-run-task>
+</monitor>
